requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
bcrypt==4.1.2
psycopg2-binary==2.9.9
requests==2.31.0
gspread==5.12.0
oauth2client==4.1.3
selenium==4.15.2
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
psutil==5.9.6
